+++ CHANGELOG.md
@@ -27,13 +27,14 @@ Released: Upcoming
 - Attention: Replaced the following functions of the class EntityDefinition with getSubTypeField:
     - getReferenceNameField
     - getReferenceEntity
+- Attention: Replaced the following functions of the class EntityDefinition with getField:
+    - getDescription
+    - getFloatStep
     - getItems
     - getValue
     - getPath
     - isUnique
-- Attention: Replaced the following functions of the class EntityDefinition with getField:
-    - getDescription
-    - getFloatStep
+    - isRequired
 - Attention: Replaced the following functions of the class EntityDefinition with setField:
     - setDescription
     - setFloatStep
@@ -41,6 +42,7 @@ Released: Upcoming
     - setValue
     - setPath
     - setUnique
+    - setRequired
 - Switched to a flag-sprites.com generated css sprite for the language flags
 - Updated dependencies:
     - "silex/silex": "~2.0"